requirements.txt
absl-py==1.0.0
attrs==21.2.0
cycler==0.11.0
fonttools==4.28.1
kiwisolver==1.3.2
matplotlib==3.5.0
mediapipe-rpi3
numpy==1.21.4
opencv-contrib-python==4.5.3.56
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyparsing==3.0.6
python-dateutil==2.8.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
msvc-runtime
